Engine: redone the rest of main/game_file code in CAGSGame
At this point it is absolutely clear that game entity should be represented by a single class. Dynamic*, Script* and *whatevername* parts should be integrated into single class that provides data and links to memory management and script system. From what I've seen in the code it should be pretty possible to derive each entity class from AGSDynamicObject and make it manage itself. A care should be taken though. I am going to investigate some more on this prior to doing aforementioned change, just to be sure.